+++ packages/@expo/cli/src/start/server/metro/instantiateMetro.ts
@@ -274,6 +274,7 @@ export async function instantiateMetroAsync(
     return originalTransformFile(
       filePath,
       pruneCustomTransformOptions(
+        projectRoot,
         filePath,
         // Clone the options so we don't mutate the original.
         {
@@ -405,6 +406,7 @@ export async function instantiateMetroAsync(
 
 // TODO: Fork the entire transform function so we can simply regex the file contents for keywords instead.
 function pruneCustomTransformOptions(
+  projectRoot: string,
   filePath: string,
   transformOptions: TransformOptions
 ): TransformOptions {
@@ -421,13 +423,20 @@ function pruneCustomTransformOptions(
     transformOptions.customTransformOptions.dom = 'true';
   }
 
-  if (
-    transformOptions.customTransformOptions?.routerRoot &&
+  const routerRoot = transformOptions.customTransformOptions?.routerRoot;
+  if (typeof routerRoot === 'string') {
+    const isRouterEntry = /\/expo-router\/_ctx/.test(filePath);
     // The router root is used all over expo-router (`process.env.EXPO_ROUTER_ABS_APP_ROOT`, `process.env.EXPO_ROUTER_APP_ROOT`) so we'll just ignore the entire package.
-    !(filePath.match(/\/expo-router\/_ctx/) || filePath.match(/\/expo-router\/build\//))
-  ) {
-    // Set to the default value.
-    transformOptions.customTransformOptions.routerRoot = 'app';
+    const isRouterModule = /\/expo-router\/build\//.test(filePath);
+    // Any page/router inside the expo-router app folder may access the `routerRoot` option to determine whether it's in the app folder
+    const isRouterRoute =
+      path.isAbsolute(filePath) && filePath.startsWith(path.resolve(projectRoot, routerRoot));
+
+    // In any other file than the above, we enforce that we mustn't use `routerRoot`, and set it to an arbitrary value here (the default)
+    // to ensure that the cache never invalidates when this value is changed
+    if (!isRouterEntry && !isRouterModule && !isRouterRoute) {
+      transformOptions.customTransformOptions!.routerRoot = 'app';
+    }
   }
 
   if (