+++ packages/expo-web-browser/ios/WebBrowserExceptions.swift
@@ -13,3 +13,10 @@ final class WebBrowserInvalidURLException: Exception {
     return "The provided URL is not valid."
   }
 }
+
+final class WebBrowserNotOpenException: Exception {
+  override var reason: String {
+    "There is no web browser to dismiss"
+  }
+}
+